pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "detraf-batimento"
version = "2.0.0"
description = "Módulo para análise automática de batimento DETRAF com CDR"
authors = [{name = "SaperX / Kevyn Martins"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12",
  "python-dotenv>=1.0",
  "PyMySQL>=1.1",
  "rich>=13.7",
  "PyYAML>=6.0"
]

[project.scripts]
detraf = "detraf.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
